web-api-helloworld/python-django-ninja/requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
click==8.3.0
Django==5.2.6
django-ninja==1.4.3
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httptools==0.6.4
idna==3.10
mysqlclient==2.2.7
packaging==25.0
pydantic==2.11.9
pydantic_core==2.33.2
PyMySQL==1.1.2
python-dotenv==1.1.1
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
